Move is_chinese_char function to almond init
